Add a basic end-to-end story for the file sources pane, factor out progress stuff
